[refactor] srs module: split to card and view
