+++ app/telegram/note.py
@@ -4,7 +4,6 @@ import logging
 from typing import Optional, Tuple, Any, List, Dict, Union
 from dataclasses import dataclass
 
-from app.srs.service import get_card
 from nachricht.llm import query_llm
 from nachricht.auth import User
 from nachricht.bus import Signal
@@ -19,11 +18,11 @@ from ..llm import (
     translate,
     detect_language,
 )
-from ..notes import language_code_by_name
+from ..notes import language_code_by_name, get_language
 from ..util import get_native_language, get_studied_language
 from ..srs import (
-    get_language,
     create_word_note,
+    get_card,
     get_notes,
     get_note,
     update_note,
@@ -101,6 +100,15 @@ class PhraseExplanationRequested(Signal):
 
 @dataclass
 class ExplanationNoteAdded(Signal):
+    """New note with an explanation of the word was created."""
+
+    note_id: int
+
+
+@dataclass
+class ExplanationNoteShown(Signal):
+    """New note with an explanation of the word was shown to the user."""
+
     note_id: int
 
 
@@ -155,7 +163,7 @@ async def get_explanation_in_native_language(ctx: Context, note: Note) -> str:
     cached_translation = note.get_option(translation_option_key)
 
     if cached_translation is not None and isinstance(cached_translation, str):
-        logger.info(
+        logger.debug(
             f"Found cached translation for note {note.id} to native language {native_language.name}."
         )
         return cached_translation
@@ -373,7 +381,7 @@ async def add_note(
     display_explanation = format_explanation(
         await get_explanation_in_native_language(ctx, note)
     )
-    return await ctx.send_message(
+    message = await ctx.send_message(
         f"{icon} *{text}* — {display_explanation}",
         reply_to=None,
         on_reaction={
@@ -384,6 +392,8 @@ async def add_note(
             "delete": NoteDeletionRequested(user_id=user.id, note_id=note.id),
         },
     )
+    bus.emit(ExplanationNoteShown(note.id), ctx=ctx)
+    return message
 
 
 @bus.on(NoteDownvoted)
